+++ src/test/java/seedu/penus/logic/LogicManagerTest.java
@@ -1,0 +1,5 @@
+package seedu.penus.logic;
+
+class LogicManagerTest {
+
+}
