+++ src/test/java/seedu/penus/logic/commands/PlanCommandTest.java
@@ -1,0 +1,5 @@
+package seedu.penus.logic.commands;
+
+class PlanCommandTest {
+
+}
