+++ src/test/java/seedu/penus/logic/commands/RemoveCommandTest.java
@@ -1,0 +1,5 @@
+package seedu.penus.logic.commands;
+
+class RemoveCommandTest {
+
+}
