+++ src/test/java/seedu/penus/logic/commands/TakenCommandTest.java
@@ -1,0 +1,5 @@
+package seedu.penus.logic.commands;
+
+class TakenCommandTest {
+
+}
