+++ src/test/java/seedu/penus/logic/parser/ParserTest.java
@@ -1,0 +1,5 @@
+package seedu.penus.logic.parser;
+
+class ParserTest {
+
+}
