+++ src/test/java/seedu/penus/logic/utils/ModuleRetrieverTest.java
@@ -1,0 +1,5 @@
+package seedu.penus.logic.utils;
+
+class ModuleRetrieverTest {
+
+}
