+++ src/test/java/seedu/penus/model/ModelManagerTest.java
@@ -1,0 +1,6 @@
+package seedu.penus.model;
+
+
+class ModelManagerTest {
+
+}
