+++ src/test/java/seedu/penus/model/ModuleTest.java
@@ -1,0 +1,6 @@
+package seedu.penus.model;
+
+
+class ModuleTest {
+
+}
