+++ src/test/java/seedu/penus/storage/FileStorageTest.java
@@ -1,0 +1,5 @@
+package seedu.penus.storage;
+
+class FileStorageTest {
+
+}
