+++ src/test/java/seedu/penus/storage/ResourceStorageTest.java
@@ -1,0 +1,5 @@
+package seedu.penus.storage;
+
+class ResourceStorageTest {
+
+}
