+++ src/test/java/seedu/penus/storage/StorageManagerTest.java
@@ -1,0 +1,5 @@
+package seedu.penus.storage;
+
+class StorageManagerTest {
+
+}
